test(tool-lane): extract overlay-rendering cluster into its own test file
Verbatim move of 11 describe blocks (50 tests, ~1,862 LOC) from
tool-lane.test.ts into a new sibling tool-lane.overlay.test.ts.

Moved blocks (getOverlay subject):
  - Spine renderer scenarios (spine geometry, flush + overlay paths)
  - ToolLane.getOverlay — MAX_OVERLAY_ROOTS sliding cap
  - ToolLane.getOverlay — childless NESTING head anchors the spine
  - ToolLane.getOverlay — suppresses ancestors already in scrollback
  - ToolLane.getOverlay — anonymous anchor for headerEmitted ancestors
  - ToolLane.getOverlay — anonymous-anchor invariant
  - ToolLane.getOverlay — headerEmitted nested skill (no grandchildren) is silent
  - ToolLane.getOverlay — depth-2 ghost row: all descendants silenced implies silence
  - ToolLane — committed spine continuity under a live skill
  - ToolLane — band/overlay descendant-row spine reconciliation
  - ToolLane.getOverlay — last-child connector closes its column (severed-spine regression)

148-test invariant holds: before = 148; after = 98 (original) + 50 (overlay) = 148.
No test added, removed, renamed, or modified — pure verbatim extraction.

Fixture strategy: makeResult() helper duplicated in the new file (4-line function,
no shared module needed). No vi.mock() or module-level beforeEach/afterEach in
either file. Each new local res() helper in individual describes moves with its block.

Snapshot hygiene: 5 Spine renderer snapshot keys migrated from
tool-lane.test.ts.snap to tool-lane.overlay.test.ts.snap.

pnpm lint: clean (tsc --noEmit).
pnpm test -- src/cli/commands/interactive/tool-lane: 5 files, 279 tests, all green.
Code under test (tool-lane.ts) untouched.
